Backend/requirements.txt
# ==============================
# Core Framework
# ==============================
fastapi==0.115.6
uvicorn[standard]==0.32.1

# ==============================
# AI / ML Components
# ==============================
google-generativeai==0.8.3
transformers==4.47.1
torch==2.5.1
sentencepiece==0.2.0

# ==============================
# Data Models & Validation
# ==============================
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# ==============================
# Domain & Network Analysis
# ==============================
python-whois==0.9.4
httpx==0.28.1

# ==============================
# Caching Layer
# ==============================
# OPTION 1: Use Redis (Production)
redis[hiredis]==5.2.1

# OPTION 2: Use FakeRedis (Testing)
# fakeredis==2.19.0

# ==============================
# Environment Management
# ==============================
python-dotenv==1.0.1

# ==============================
# Async Utilities & Reliability
# ==============================
asyncio-throttle==1.0.2
tenacity==9.0.0
aiofiles==24.1.0
slowapi==0.1.9

# ==============================
# Logging
# ==============================
loguru==0.7.3
